unify 2.12/2.13 SparkeyMap
